+++ pyhidra/__init__.py
@@ -1,13 +1,15 @@
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"
 
 # Expose API
 from .ghidra import run_script, start, open_program
 from .script import get_current_interpreter
 from .launcher import DeferredPyhidraLauncher, HeadlessPyhidraLauncher, GuiPyhidraLauncher
+from .version import ExtensionDetails
 
 
 __all__ = [
     "run_script", "start", "open_program",
-    "DeferredPyhidraLauncher", "HeadlessPyhidraLauncher", "GuiPyhidraLauncher"
+    "DeferredPyhidraLauncher", "HeadlessPyhidraLauncher", "GuiPyhidraLauncher",
+    "ExtensionDetails",
 ]