+++ pyhidra/java/plugin/__init__.py
@@ -1,4 +1,3 @@
 """
 Internal use only
 """
-from .handler import install